% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_pathways}
\alias{demo_pathways}
\title{Synthetic demonstration pathway collection}
\format{
named list of 6 character vectors.
}
\usage{
data(demo_pathways)
}
\description{
Six disjoint 30-gene pathways over the \code{demo_z} gene vocabulary.
Members of PATHWAY_A, PATHWAY_C, and PATHWAY_E are among the perturbed
genes; see \code{demo_z}.
}
\keyword{datasets}
