% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2path-package.R
\docType{package}
\name{path2path-package}
\alias{path2path-package}
\title{path2path: directional pathway-by-pathway maps from Perturb-seq}
\description{
The package converts a perturbation-by-gene matrix of signed
differential-expression statistics into a directional pathway-by-pathway
map using two successive preranked gene set enrichment analyses. It also
provides pathway-overlap safeguards, two routes for significance
assessment, plotting functions, and conversion to and from
`SummarizedExperiment` objects.
}
\details{
See `vignette("path2path")` for an introduction and worked example.
}
\references{
Subramanian A, et al. (2005). Gene set enrichment analysis: a
knowledge-based approach for interpreting genome-wide expression profiles.
Proceedings of the National Academy of Sciences 102:15545-15550.
doi:10.1073/pnas.0506580102.

Korotkevich G, Sukhov V, Sergushichev A (2019). Fast gene set enrichment
analysis. bioRxiv. doi:10.1101/060012.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jli-stat/path2path}
  \item Report bugs at \url{https://github.com/jli-stat/path2path/issues}
}

}
\author{
Jun Li
}
\keyword{internal}
